requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
environ==1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
oauthlib==3.2.0
openpyxl==3.0.10
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.10
uvicorn==0.18.2
